refactor: :recycle: Renombrado del modulo de aplicaciones
Renombrado del modulo de aplicaciones
