+++ react-starter/src/modules/aplicaciones/config/ColumnasListado.jsx
