+++ react-starter/src/modules/aplicaciones/pages/index.js
