+++ react-starter/src/modules/index.js
@@ -1,4 +1,4 @@
 export * from "./consejodegobierno"
 export * from "./dashboard"
-export * from "./listadoaplicaciones"
+export * from "./aplicaciones"
 export * from "./utilidades"
